refactor view-elm logic
