+++ src/debug/stepGranularityToggle.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { Commands } from '../commands/commands';
-import { getActiveSplitDebugHook } from '../commands/view-elm';
 import * as log from '../log-services/logger';
+import { AstSplitSessionManager } from '../views/astSplitSession';
 
 type Granularity = 'cql' | 'ast';
 const sessionGranularity = new WeakMap<vscode.DebugSession, Granularity>();
@@ -34,7 +34,7 @@ export function activateStepGranularityToggle(context: vscode.ExtensionContext) 
       await s.customRequest('setStepGranularity', { granularity: next });
       sessionGranularity.set(s, next);
       refresh(s);
-      if (next === 'ast' && !getActiveSplitDebugHook()) {
+      if (next === 'ast' && !AstSplitSessionManager.getActiveSession()) {
         const cqlEditor = vscode.window.visibleTextEditors.find(
           e => e.document.uri.fsPath.toLowerCase().endsWith('.cql'),
         );
